Seperated actions into seperate fragments
