+++ app/src/main/java/com/thing/now/listener/OnSwitchListener.kt
@@ -1,0 +1,6 @@
+package com.thing.now.listener
+
+
+interface OnSwitchListener {
+    fun onSwitch(id: Int)
+}
