+++ app/src/main/res/layout/fragment_acknowledgements.xml
@@ -1,15 +1,60 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              tools:context=".fragment.AcknowledgementsFragment"
-              android:orientation="vertical"
->
+<ScrollView
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".fragment.AcknowledgementsFragment"
+        android:padding="@dimen/margin_medium">
+    <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
 
-    <TextView
-            android:text="vale"
-            android:layout_width="wrap_content"
-              android:layout_height="wrap_content"/>
+        <TextView
+                android:text="@string/app_license_header"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"/>
 
-</LinearLayout>
+        <android.support.v4.widget.NestedScrollView
+                android:padding="@dimen/margin_small"
+                android:background="@color/colorNotNow"
+                android:layout_marginTop="@dimen/margin_small"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/peek_height">
+            <TextView
+                    android:text="@string/gpl2_license"
+                    android:fontFamily="monospace-"
+                    android:id="@+id/appLicenseTextView"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"/>
+        </android.support.v4.widget.NestedScrollView>
+
+        <TextView
+                android:layout_marginTop="@dimen/margin_medium"
+                android:text="@string/ack_os_used_header"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"/>
+
+        <LinearLayout
+                android:id="@+id/oslNameContainer"
+                android:layout_marginTop="@dimen/margin_small"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical"/>
+
+        <TextView
+                android:layout_marginTop="@dimen/margin_medium"
+                android:text="@string/ack_icons_used_header"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"/>
+
+        <LinearLayout
+                android:id="@+id/iconAckContainer"
+                android:layout_marginTop="@dimen/margin_small"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical"/>
+
+    </LinearLayout>
+</ScrollView>
