+++ app/src/main/res/values/dimens.xml
@@ -14,9 +14,9 @@
     <dimen name="notification_dot_size">16dp</dimen>
     <dimen name="user_image_size">48dp</dimen>
 
-    <dimen name="elevation_small">4dp</dimen>
-    <dimen name="elevation_medium">6dp</dimen>
-    <dimen name="elevation_large">8dp</dimen>
+    <dimen name="elevation_small">2dp</dimen>
+    <dimen name="elevation_medium">4dp</dimen>
+    <dimen name="elevation_large">6dp</dimen>
 
     <dimen name="radius_small">4dp</dimen>
     <dimen name="radius_medium">8dp</dimen>
