+++ app/src/main/res/values/icon_names.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="icon_names">Flemingo by Flemingo</string-array>
+</resources>
